+++ src/main/java/me/unariginal/novaraids/commands/RaidCommands.java
@@ -412,12 +412,12 @@ public class RaidCommands {
     private int modInfo(CommandContext<ServerCommandSource> ctx) {
         ServerPlayerEntity player = ctx.getSource().getPlayer();
         if (player != null) {
-            player.sendMessage(TextUtils.deserialize("<gray><st><b><i>--------- <reset><red>Nova Raids <reset><gray><st><b><i>---------"));
+            player.sendMessage(TextUtils.deserialize("<gray><st><b><i>---------<reset><red> Nova Raids <reset><gray><st><b><i>---------"));
             player.sendMessage(TextUtils.deserialize("<gray><b>Author: <reset><white>Unariginal <i>(Ariginal)"));
             player.sendMessage(TextUtils.deserialize("<gray><b>Version: <reset><white>Beta v0.3.1"));
             player.sendMessage(TextUtils.deserialize("<gray><b><i><u><click:open_url:https://github.com/Unariginal/NovaRaids>Source"));
             player.sendMessage(TextUtils.deserialize("<gray><b><i><u><click:open_url:https://github.com/Unariginal/NovaRaids/wiki>Wiki"));
-            player.sendMessage(TextUtils.deserialize("<gray><st><b><i>----------------------------"));
+            player.sendMessage(TextUtils.deserialize("<gray><st><b><i>---------------------------"));
         }
         return 1;
     }